// src/Models/EmployeeTask.tsx
export interface EmployeeTask {
    id: number;
    employeeId: number;
    taskId: number;
    projectId: number;
    startDate: Date;
    endDate: Date;
    estTime: number;
    actTime: number;
    weekEndingDate: Date;
    status: string;
    priority: string;
    percentage: number;
    name:string;
    description:string;
    estStartDate:Date;
    estEndDate:Date;
    createdDate:Date;
    createdBy:string;
    updatedDate:Date;
    updatedBy:string;
    taskDescription:string;
    taskType:string;
    projectName:string;
  }












  